Squashed 'airdcpp-webapi/' changes from 567e70d..41419f6
41419f6 API: various table fixes, serialize more information about filelists
31101ce Use VERIFY state also with NMDC hubs (fixes "password" state in API)
50a4b73 Fix a few compiler warnings with the 32 bit build

git-subtree-dir: airdcpp-webapi
git-subtree-split: 41419f62e15cfd1c34df8d79bd68fd7e1f2363e3

+++ api/ApiModule.cpp
@@ -66,7 +66,7 @@ namespace webserver {
 			return false;
 		}
 
-		for (auto i = 0; i < params.size(); i++) {
+		for (auto i = 0; i < static_cast<int>(params.size()); i++) {
 			if (!params[i].match(aRequestParams[i])) {
 				return false;
 			}
